+++ src/lib/scriptengine/rsscript_test.rs
@@ -12,6 +12,16 @@ fn get_script_runner_default() {
 
 #[test]
 #[ignore]
+fn get_script_runner_rust_script() {
+    envmnt::set("CARGO_MAKE_RUST_SCRIPT_PROVIDER", "rust-script");
+    let provider = get_script_runner();
+    envmnt::remove("CARGO_MAKE_RUST_SCRIPT_PROVIDER");
+
+    assert_eq!(provider, ScriptRunner::RustScript);
+}
+
+#[test]
+#[ignore]
 fn get_script_runner_cargo_script() {
     envmnt::set("CARGO_MAKE_RUST_SCRIPT_PROVIDER", "cargo-script");
     let provider = get_script_runner();
@@ -93,6 +103,72 @@ fn execute_default_runtime_panic() {
 fn execute_default_runtime_panic_no_validate() {
     if test::should_test(false) {
         envmnt::remove("CARGO_MAKE_RUST_SCRIPT_PROVIDER");
+
+        execute(
+            &vec!["fn main() {panic!(\"error\");}".to_string()],
+            &vec![],
+            false,
+        );
+
+        envmnt::remove("CARGO_MAKE_RUST_SCRIPT_PROVIDER");
+    }
+}
+
+#[test]
+#[ignore]
+fn execute_rust_script_valid() {
+    if test::should_test(false) {
+        envmnt::set("CARGO_MAKE_RUST_SCRIPT_PROVIDER", "rust-script");
+
+        execute(
+            &vec!["fn main() {println!(\"test\");}".to_string()],
+            &vec![],
+            true,
+        );
+
+        envmnt::remove("CARGO_MAKE_RUST_SCRIPT_PROVIDER");
+    }
+}
+
+#[test]
+#[ignore]
+#[should_panic]
+fn execute_rust_script_not_compile() {
+    if test::should_test(true) {
+        envmnt::set("CARGO_MAKE_RUST_SCRIPT_PROVIDER", "rust-script");
+
+        execute(
+            &vec!["fn main() {donotcompile();}".to_string()],
+            &vec![],
+            true,
+        );
+
+        envmnt::remove("CARGO_MAKE_RUST_SCRIPT_PROVIDER");
+    }
+}
+
+#[test]
+#[ignore]
+#[should_panic]
+fn execute_rust_script_runtime_panic() {
+    if test::should_test(true) {
+        envmnt::set("CARGO_MAKE_RUST_SCRIPT_PROVIDER", "rust-script");
+
+        execute(
+            &vec!["fn main() {panic!(\"error\");}".to_string()],
+            &vec![],
+            true,
+        );
+
+        envmnt::remove("CARGO_MAKE_RUST_SCRIPT_PROVIDER");
+    }
+}
+
+#[test]
+#[ignore]
+fn execute_rust_script_runtime_panic_no_validate() {
+    if test::should_test(false) {
+        envmnt::set("CARGO_MAKE_RUST_SCRIPT_PROVIDER", "rust-script");
 
         execute(
             &vec!["fn main() {panic!(\"error\");}".to_string()],