+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v0.32.13
 
+* Enhancement: Support rust-script as an additional rust runner #516
 * Fix: cm_run_task now also invokes dependencies of the target task #519
 * Fix: Env vars provided on cli are not positioned first when evaluated #520